Move getModuleByAddress to lookup module
